WIP: This is mess
- I don't know what kind of mess I have created, but I am anything but
happy right now. What the shit happened I don't know! And, why it
happened ? I don't know. But, here we are. So enjoy dear reader!!!!

+++ cmd/mygit/clone_test.go
@@ -64,7 +64,7 @@ func Test_ReadPackFileHeader(t *testing.T) {
 
 	testCases := [...]struct {
 		input            []byte
-		outputSize       int
+		outputSize       uint64
 		outputObjectType ObjectType
 	}{
 		{
